# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.CreateUserReq:
    properties:
      date_of_birth:
        example: "1990-01-01"
        type: string
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
    required:
    - date_of_birth
    - email
    - name
    type: object
  models.UpdateUserReq:
    properties:
      email:
        example: jane@example.com
        type: string
      name:
        example: Jane Doe
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      date_of_birth:
        type: string
      email:
        type: string
      group:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: mohammadaliabudalou@example.com
    name: Mohammad Ali Abu-Dalou
  description: REST API in Go (Gin + GORM) with automatic group assignment.
  title: Backend Task API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Returns a list of users, optionally filtered by group using query
        parameter (e.g. adult-1, senior-2).
      parameters:
      - description: Group name
        in: query
        name: group
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates new users and assigns them to groups automatically (up
        to 3 per group).
      parameters:
      - description: User info array
        in: body
        name: users
        required: true
        schema:
          items:
            $ref: '#/definitions/models.CreateUserReq'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Create one or more users
      tags:
      - users
  /users/{id}:
    get:
      description: Retrieve a user by their unique UUID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update user name and/or email by ID (group cannot be updated manually).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request or ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update a user
      tags:
      - users
swagger: "2.0"